+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ##[TBD]
 * Expose extra deviceInfo.
+* Ignore sign in status on account removal from ODSP cache #1541
 * Allow adding additional query string parameters when opening the browser. (#1539)
 
 ## [1.2.3]